Fixes bug with superstructure db export
The production flows were not listed in the scenario difference file.
Also, in the case of losses, those were replacing the production flow values, leading to wrong results.
Now losses are subtracted from the production flow value (which is typically 1).
For example, for electricity markets, the output flow is no longer 1 kWh but 1 - loss.